+++ apps/frontend/src/libs/components/app/app.tsx
@@ -1,4 +1,4 @@
-import { Loader, RouterOutlet } from "~/libs/components/components.js";
+import { RouterOutlet } from "~/libs/components/components.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -8,6 +8,8 @@ import {
 } from "~/libs/hooks/hooks.js";
 import { actions as appActions } from "~/libs/slices/app/app.js";
 import { actions as authActions } from "~/modules/auth/auth.js";
+
+import { Loader } from "../loader/loader.js";
 
 const App: React.FC = () => {
 	const dispatch = useAppDispatch();