+++ packages/shared/src/libs/types/pagination-request-dto.ts
@@ -1,0 +1,6 @@
+type PaginationRequestDto = {
+	count: number;
+	page: number;
+};
+
+export { type PaginationRequestDto };