+++ packages/shared/src/modules/user-courses/users-courses.ts
@@ -1,2 +1,5 @@
 export { UserCoursesApiPath } from "./libs/enums/enums.js";
-export { userIdParameter as userIdParameterValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+export {
+	userCourseGetAllQuery as userCourseGetAllQueryValidationSchema,
+	userIdParameter as userIdParameterValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";